requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1751547525079/work
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bleach==6.2.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
comm==0.2.2
contourpy @ file:///croot/contourpy_1732540045555/work
cryptography==45.0.5
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dashscope==1.23.8
datasets==1.18.4
debugpy==1.8.14
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
decord==0.6.0
defusedxml==0.7.1
diffusers==0.35.2
dill==0.3.7
diskcache==5.6.3
entmax==1.3
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fonttools @ file:///croot/fonttools_1737039080035/work
fqdn==1.5.1
frozenlist==1.6.0
fschat==0.2.36
fsspec==2023.6.0
googledrivedownloader==1.1.0
graphviz==0.21
groovy==0.1.2
grpcio @ file:///croot/grpc-split_1742490263735/work
gudhi==3.11.0
h11==0.16.0
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
lark==1.2.2
latex2mathml==3.77.0
levenshtein==0.21.1
lit==18.1.8
litellm==1.74.4
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1762201306589/work
markdown-it-py==3.0.0
markupsafe==2.0.1
matplotlib==3.10.0
mcp==1.12.0
mdurl==0.1.2
mistune==3.1.3
modelscope==1.30.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.15
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkit==11.1.post1
networkx==3.4.2
nh3==0.2.21
ninja==1.11.1.4
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
overrides==7.7.0
packaging @ file:///croot/packaging_1734472117206/work
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.10.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp310-abi3-linux_x86_64.whl#sha256=e55acfd476c7a1d8544b98df45c7d8e9c425b9caf0abb1f7e7cfa87c2e017bd1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==14.0.2
pycountry==24.6.1
pycparser==2.22
pydantic==1.10.22
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing @ file:///croot/pyparsing_1731445506121/work
pypdfium2==4.30.1
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1753427276959/work/pyqt_sip
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-louvain==0.16
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
qwen-agent==0.0.27
rank-bm25==0.2.2
ray==2.48.0
rdflib==7.1.4
referencing==0.36.2
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.24.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
sip @ file:///croot/sip_1738856193618/work
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sparse==0.17.0
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tensorboard_1752825441/work/tensorboard-2.20.0-py3-none-any.whl#sha256=9dc9f978cb84c0723acf9a345d96c184f0293d18f166bb8d59ee098e6cfaaba6
tensorboard_data_server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1759413119425/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=8ec18459e223309f6d0d987e0256a30babd90bf53a779c714797a68b70ed1ad8
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.15.2
tomli==2.2.1
TopoNetX==0.2.0
torch==2.1.0+cu118
torch-cluster==1.6.3+pt21cu118
torch-geometric==2.3.1
torch-scatter==2.1.2+pt21cu118
torch-sparse==0.6.18+pt21cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
trimesh==4.6.8
triton==2.1.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.0
tzdata==2025.2
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
uri-template==1.3.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.1.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
widgetsnbextension==4.0.14
xformers==0.0.22
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
